requirements.txt
numpy
scipy
pyroomacoustics
soundfile
librosa
PyYAML
pandas
openai
